pyproject.toml
[project]
name = "code-review-env"
version = "1.0.0"
description = "OpenEnv environment for code review and security audit training"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "websockets>=11.0",
    "openenv-core>=0.2.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio"]

[project.scripts]
serve = "server.app:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"
